requirements.txt
discord.py[voice]
dice